--- a/app/models/host/managed.rb
+++ b/app/models/host/managed.rb
@@ -846,7 +846,7 @@ class Host::Managed < Host::Base
 
   def refresh_global_status!
     refresh_global_status
-    save!
+    save!(:validate => false)
   end
 
   def refresh_statuses