@dLobatog 2016-04-22
How about running `.valid?` first and raising a `Foreman::Exception` if it's not valid? I think that'd be a good compromise. I agree invalid statuses should not show up, but if they show up, at least they can go to the wiki or report something. Otherwise it's a shot in the dark for us to know what happened.