@ares 2016-04-25
> How about running .valid? first and raising a Foreman::Exception if it's not valid? I think that'd be a good compromise.

Sounds reasonable. The only reason why I hesitate is that there are multiple ways in Foreman that can make host invalid, e.g. unmanaged host -> managed host transition. So maybe we should fix these issues first? This can take a while though.