@dLobatog 2016-04-28
> Sounds reasonable. The only reason why I hesitate is that there are multiple ways in Foreman that can make host invalid, e.g. unmanaged host -> managed host transition.

'host invalid' you meant 'host status', right? If you can list some of the issues that cause this, maybe I can help you sending PRs with the fixes too.