@ares 2016-04-28
No, I mean host invalid. Just create Host by incoming report/facts, then in UI go to edit form, click Manage host button. Now when you call `.valid?` on it, it will most likely return false.