@tbrisker 2016-07-25
Is there any situation when the status update can change any host attributes other then the global_status? if not, I don't see any point in running all the validations on every status update. I would even consider updating it without invoking callbacks at all - if i understand correctly this is executed every time a report is received, so if possible it would be best to only update the column in the database (unless some callbacks are needed)