@ares 2016-07-27
> Is there any situation when the status update can change any host attributes other then the global_status?

Yes, the status should not change any other host attribute. And since global_status does not have any specific validation it should be fine.

> if i understand correctly this is executed every time a report is received

correct, mostly because of `reported_at` update, status does not change that frequently I guess

> so if possible it would be best to only update the column in the database

I'd like to avoid this unless we're sure that it would have huge performance benefit. I think one update of an active record is not the slowest part of configuration report importing.

Thanks @tbrisker for your thoughts, I think the fix is appropriate. Let's hope users would notice elsewhere if the host was invalid. If we get https://github.com/theforeman/foreman/pull/3670 in at some point, it should be visible from the dashboard widget too.