@tbrisker 2016-07-27
@ares (tl;dr: there is a much faster way to do it, see near the bottom)

using activerecord causes all the callbacks on the host to be called, here is the list of callbacks registered on Host::Managed, without any plugins:
```ruby
[23] pry(main)> Host::Managed._save_callbacks.map(&:filter)
=> [:autosave_associated_records_for_model,
 :autosave_associated_records_for_location,
 :autosave_associated_records_for_organization,
 :autosave_associated_records_for_search_users,
 :autosave_associated_records_for_hostgroup,
 :autosave_associated_records_for_owner,
 :autosave_associated_records_for_compute_resource,
 :autosave_associated_records_for_image,
 :clear_data_on_build,
 :clear_puppetinfo,
 :autosave_associated_records_for_architecture,
 :autosave_associated_records_for_environment,
 :autosave_associated_records_for_operatingsystem,
 :autosave_associated_records_for_medium,
 :autosave_associated_records_for_ptable,
 :autosave_associated_records_for_puppet_proxy,
 :autosave_associated_records_for_puppet_ca_proxy,
 :autosave_associated_records_for_realm,
 :autosave_associated_records_for_compute_profile,
 :check_puppet_ca_proxy_is_required?,
 :crypt_root_pass,
 :set_lookup_value_matcher,
 :before_save_collection_association,
 :around_save_orchestration]
[24] pry(main)> Host::Managed._commit_callbacks.map(&:filter)
=> [:build_hooks, :post_commit]
[25] pry(main)> Host::Managed._update_callbacks.map(&:filter)
=> [:autosave_associated_records_for_audits,
 :autosave_associated_records_for_associated_audits,
 :autosave_associated_records_for_lookup_values,
 :autosave_associated_records_for_group_puppetclasses,
 :autosave_associated_records_for_config_group_classes,
 :autosave_associated_records_for_config_groups,
 :autosave_associated_records_for_host_config_groups,
 :autosave_associated_records_for_configuration_status_object,
 :autosave_associated_records_for_host_statuses,
 :autosave_associated_records_for_last_report_object,
 :autosave_associated_records_for_reports,
 :autosave_associated_records_for_puppetclasses,
 :autosave_associated_records_for_host_classes,
 :autosave_associated_records_for_host_parameters,
 :autosave_associated_records_for_parameters,
 :autosave_associated_records_for_token,
 :autosave_associated_records_for_search_parameters,
 :autosave_associated_records_for_interfaces,
 :autosave_associated_records_for_subnet6,
 :autosave_associated_records_for_subnet,
 :autosave_associated_records_for_domain,
 :autosave_associated_records_for_provision_interface,
 :autosave_associated_records_for_primary_interface,
 :autosave_associated_records_for_fact_names,
 :autosave_associated_records_for_fact_values,
 :audit_update]
```

However, surprisingly this is actually slightly faster then using just `update_column` when there are no changes to the status:
```
[43] pry(main)> Benchmark.ips do |x|
[43] pry(main)*   x.report("update_column") {h.update_column(:global_status, h.build_global_status.status)}  
[43] pry(main)*   x.report("new refresh_global_status") {h.global_status=h.build_global_status.status;h.save!(:validate=>false)}  
[43] pry(main)*   x.report("current refresh_global_status") {h.refresh_global_status!}  
[43] pry(main)* end  
Warming up --------------------------------------
       update_column    24.000  i/100ms
new refresh_global_status
                        24.000  i/100ms
current refresh_global_status
                         5.000  i/100ms
Calculating -------------------------------------
       update_column    245.009  (±10.2%) i/s -      1.200k
new refresh_global_status
                        300.610  (±33.3%) i/s -      1.200k in   5.278382s
current refresh_global_status
                         65.302  (±30.6%) i/s -    265.000 
```

I'm guessing this result is because activerecord only commits to DB if there are changes, which there aren't if the status does not change, but when using `update_column` there is always a transaction even when there is no actual change. 
Which led me to think about only updating the column if needed, with the following results:

```
[58] pry(main)> Benchmark.ips do |x|
[58] pry(main)*   x.report("current refresh_global_status") {h.refresh_global_status!}  [58] pry(main)*   x.report("new refresh_global_status") {h.global_status=h.build_global_status.status;h.save!(:validate=>false)}  
[58] pry(main)*   x.report("update_column") {h.update_column(:global_status, h.build_global_status.status)}  [58] pry(main)*   x.report("conditional update_column") {status=h.build_global_status.status;h.update_column(:global_status, status) unless h.global_status==status}  
[58] pry(main)*   x.compare!
[58] pry(main)* end  
Warming up --------------------------------------
current refresh_global_status
                         5.000  i/100ms
new refresh_global_status
                        23.000  i/100ms
       update_column    24.000  i/100ms
conditional update_column
                       362.000  i/100ms
Calculating -------------------------------------
current refresh_global_status
                         66.260  (±30.2%) i/s -    265.000 
new refresh_global_status
                        306.199  (±32.3%) i/s -      1.150k
       update_column    243.609  (±11.1%) i/s -      1.200k
conditional update_column
                          9.606k (±69.8%) i/s -     18.824k

Comparison:
conditional update_column:     9605.9 i/s
new refresh_global_status:      306.2 i/s - 31.37x slower
       update_column:      243.6 i/s - 39.43x slower
current refresh_global_status:       66.3 i/s - 144.97x slower
```

A X30 improvement over disabling validations and almost X150(!) improvement compared to our current situation when no update is needed! 
Since this code is executed for each incoming report, which could be in the tens of thousands or more every half hour, I think unless there is a good reason we want callbacks to trigger here this improvement could be significant for large setups.
Thanks and sorry for the very long comment.